// src/types/database.d.ts
import type {
  ClipboardContentType,
  ReadClipboardItemUnion,
} from "tauri-plugin-clipboard-x-api";
import type { LiteralUnion } from "type-fest";

export type DatabaseSchemaHistorySubtype = "url" | "email" | "color" | "path";

export type DatabaseSchemaHistory<
  T extends ClipboardContentType = ClipboardContentType,
> = ReadClipboardItemUnion<T> & {
  id: string;
  group: DatabaseSchemaGroupId;
  search: string;
  favorite: boolean;
  createTime: string;
  note?: string;
  edited?: boolean;
  source?: string;
  sourcePath?: string;
  tags?: string;
  subtype?: DatabaseSchemaHistorySubtype;
};

export type DatabaseSchemaGroupId = LiteralUnion<
  | "all"
  | "sync"
  | "text"
  | "image"
  | "files"
  | "favorite"
  | "rtf"
  | "url"
  | "color"
  | "path"
  | "email",
  string
>;

export interface DatabaseSchemaGroup {
  id: DatabaseSchemaGroupId;
  name: string;
  createTime?: string;
}

export interface DatabaseSchema {
  history: DatabaseSchemaHistory;
  group: DatabaseSchemaGroup;
}